refactor: Moved engine wrapper to src/
